Consolidate styling and implement ground work for transformer mode
Now most of the design is ready. We have a fringe, a main content and a sidebar.

Text is wrapped in a copy-wrapper div both to allow correct spacing between
textual elements inside of a grid with automatically generated rows, and to
allow us to easily hide the explanations, literally turning on a micro
playground version of the website, only with the text fields.

The website consists of two grids. The main one has three columns, where only
the middle one is content. The first and second are spacing columns that shrink
down to 0, giving priority to the middle column when in smaller view ports.

The second grid is where the main content lives. It also consists of three
columns, but with very different roles. The first one contains a fringe where we
put only our input labels; the second is where the paragraphs and input fields
live; and the third is something like a sidebar where we currently have our logo.